refactor: Move fixtures to configtest.py for Unit Tests and some Acceptation and Integration Tests. Remove useless fixtures imports in tests. docs: Add licence information in __init__ file
