+++ tests/acceptation/conftest.py
@@ -26,36 +26,38 @@
 # CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
 # OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 # OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-# pylint: unused-import
-import os
+from multiprocessing import active_children
+
 import pytest
 
-from powerapi.cli import ConfigValidator
+from tests.utils.db.mongo import MONGO_URI, MONGO_DATABASE_NAME, MONGO_OUTPUT_COLLECTION_NAME
 
 
 @pytest.fixture
-def create_empty_files_from_config(invalid_csv_io_stream_config):
-    for key, input_config in invalid_csv_io_stream_config['input'].items():
-        if input_config['type'] == 'csv':
-            list_of_files = input_config['files'].split(",")
-            for file_str in list_of_files:
-                if os.path.isfile(file_str) is False:
-                    with open(file_str, 'w') as file:
-                        file.close()
-
-    yield
-
-    for key, input_config in invalid_csv_io_stream_config['input'].items():
-        if input_config['type'] == 'csv':
-            list_of_files = input_config['files']
-            for file_str in list_of_files:
-                if os.path.isfile(file_str):
-                    os.remove(file_str)
+def shutdown_system():
+    """
+    Shutdown the actor system, i.e., all actors are killed
+    """
+    yield None
+    active = active_children()
+    for child in active:
+        child.kill()
 
 
-def test_invalid_config_stream_and_csv_input(invalid_csv_io_stream_config):
-    assert ConfigValidator.validate(invalid_csv_io_stream_config) is False
-
-
-def test_valid_config_postmortem_csv_input(create_empty_files_from_config, csv_io_postmortem_config):
-    assert ConfigValidator.validate(csv_io_postmortem_config)
+@pytest.fixture()
+def socket_info_config():
+    """
+    Return a configuration with socket as input and mongo as output
+    """
+    return {'verbose': True,
+            'stream': False,
+            'input': {'test_puller': {'type': 'socket',
+                                      'port': 0,
+                                      'model': 'HWPCReport'}},
+            'output': {'test_pusher': {'type': 'mongodb',
+                                       'uri': MONGO_URI,
+                                       'db': MONGO_DATABASE_NAME,
+                                       'model': 'PowerReport',
+                                       'max_buffer_size': 0,
+                                       'collection': MONGO_OUTPUT_COLLECTION_NAME}},
+            }
