+++ tests/integration/test_dispatcher.py
@@ -38,8 +38,6 @@ from powerapi.dispatch_rule import HWPCDispatchRule, HWPCDepthLevel
 from tests.utils.dummy_actor import DummyActor
 from tests.utils.report.hwpc import gen_HWPCReports
 from tests.unit.actor.abstract_test_actor import recv_from_pipe
-# noinspection PyUnresolvedReferences
-from tests.utils.unit import shutdown_system
 
 PUSHER_NAME_POWER_REPORT = 'fake_pusher_power'
 REPORT_TYPE_TO_BE_SENT = PowerReport
