+++ tests/unit/puller/simple/test_simple_puller_actor.py
@@ -27,7 +27,7 @@
 # OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 # OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-# pylint: disable=arguments-differ,unused-argument,no-self-use,redefined-outer-name,unused-import
+# pylint: disable=arguments-differ,no-self-use,unused-argument
 
 from time import sleep
 
@@ -41,9 +41,6 @@ from powerapi.puller.simple.simple_puller_actor import SimplePullerActor
 from tests.utils.dummy_actor import DummyActor
 
 from tests.unit.actor.abstract_test_actor import AbstractTestActor, recv_from_pipe, is_actor_alive
-
-# noinspection PyUnresolvedReferences
-from tests.utils.unit import shutdown_system
 
 
 NUMBER_OF_REPORTS_TO_SEND = 100
