+++ tests/unit/pusher/simple/test_simple_pusher_actor.py
@@ -27,7 +27,7 @@
 # OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 # OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-# pylint: disable=redefined-outer-name,unused-argument,unused-import,no-self-use
+# pylint: disable=unused-argument,no-self-use
 
 from time import sleep
 
@@ -36,8 +36,6 @@ import pytest
 from powerapi.message import PoisonPillMessage, GetReceivedReportsSimplePusherMessage
 from powerapi.pusher.simple.simple_pusher_actor import SimplePusherActor
 from powerapi.report import PowerReport, HWPCReport
-# noinspection PyUnresolvedReferences
-from tests.utils.unit import shutdown_system
 from tests.unit.actor.abstract_test_actor import is_actor_alive, AbstractTestActor
